Limit grpc_json_transcoder source files to one logger ID
